+++ src/items.lua
@@ -1,5 +1,4 @@
 local _, Addon = ...
-local Container = Addon:GetModule("Container")
 local E = Addon:GetModule("Events")
 local EquipmentSetsCache = Addon:GetModule("EquipmentSetsCache")
 local EventManager = Addon:GetModule("EventManager")
@@ -22,7 +21,7 @@ do
   local t = {}
 
   function getContainerItem(bag, slot)
-    local item = Container.GetContainerItemInfo(bag, slot)
+    local item = C_Container.GetContainerItemInfo(bag, slot)
     if type(item) ~= "table" then return nil end
 
     for k in pairs(t) do t[k] = nil end
@@ -69,7 +68,7 @@ end
 
 local function iterateBags()
   local bag, slot = BACKPACK_CONTAINER, 0
-  local numSlots = Container.GetContainerNumSlots(bag)
+  local numSlots = C_Container.GetContainerNumSlots(bag)
 
   return function()
     slot = slot + 1
@@ -81,11 +80,11 @@ local function iterateBags()
       repeat
         bag = bag + 1
         if bag > NUM_BAG_SLOTS then return nil end
-        numSlots = Container.GetContainerNumSlots(bag)
+        numSlots = C_Container.GetContainerNumSlots(bag)
       until numSlots > 0
     end
 
-    return bag, slot, Container.GetContainerItemID(bag, slot)
+    return bag, slot, C_Container.GetContainerItemID(bag, slot)
   end
 end
 
@@ -140,6 +139,11 @@ end
 -- Bags
 -- ============================================================================
 
+function Items:GetFreshItem(bag, slot)
+  EquipmentSetsCache:Refresh()
+  return getItem(bag, slot)
+end
+
 function Items:GetItem(bag, slot)
   for _, item in pairs(self.cache) do
     if item.bag == bag and item.slot == slot then
@@ -164,11 +168,11 @@ function Items:GetItems(items)
 end
 
 function Items:IsBagSlotEmpty(bag, slot)
-  return Container.GetContainerItemID(bag, slot) == nil
+  return C_Container.GetContainerItemID(bag, slot) == nil
 end
 
 function Items:IsItemStillInBags(item)
-  return item.id == Container.GetContainerItemID(item.bag, item.slot)
+  return item.id == C_Container.GetContainerItemID(item.bag, item.slot)
 end
 
 function Items:IsItemLocked(item)
@@ -207,7 +211,7 @@ function Items:IsItemDestroyable(item)
 end
 
 function Items:IsItemRefundable(item)
-  local refundTimeRemaining = select(3, Container.GetContainerItemPurchaseInfo(item.bag, item.slot, false))
+  local refundTimeRemaining = select(3, C_Container.GetContainerItemPurchaseInfo(item.bag, item.slot, false))
   return refundTimeRemaining and refundTimeRemaining > 0
 end
 
@@ -224,13 +228,17 @@ do -- Items:IsItemEquipment()
 
     if item.classId == Enum.ItemClass.Armor then
       if invTypeExceptions[item.invType] then return true end
-      return not (item.subclassId == Enum.ItemArmorSubclass.Generic or
-          item.subclassId == Enum.ItemArmorSubclass.Cosmetic)
+      return not (
+        item.subclassId == Enum.ItemArmorSubclass.Generic or
+        item.subclassId == Enum.ItemArmorSubclass.Cosmetic
+      )
     end
 
     if item.classId == Enum.ItemClass.Weapon then
-      return not (item.subclassId == Enum.ItemWeaponSubclass.Generic or
-          item.subclassId == Enum.ItemWeaponSubclass.Fishingpole)
+      return not (
+        item.subclassId == Enum.ItemWeaponSubclass.Generic or
+        item.subclassId == Enum.ItemWeaponSubclass.Fishingpole
+      )
     end
 
     return false